infoschema: move perfschema to infoschema directory, tiny refactor
* move perfschema package to github.com/pingcap/tidb/infoschema/perfschema
* introduce infoschema.VirtualTable and simplify perfschema.perfSchemaTable
